fix: fix main page requests
